backend/requirements.txt
fastapi
uvicorn
sqlalchemy
groq
httpx
python-dotenv
cryptography
google-genai